sw/python/requirements.txt
matplotlib==3.4.1
numpy==1.19.5
pylink-square==0.10.0
pyserial==3.4
scipy==1.10.0